Fixed parsing of comments
I misunderstood what we get from comment/list endpoint. It seems that we do not
receive a list ready for display and some processing on the client is assumed.
It seems we get a list of top level comments and then children and children of
children.

The response is partially ordered - the relative order of commits shall be
preserved but the grouping of parent->child commits is done in the client.