Widget: enable Siri Suggestions on the intent
Donate intent when the user opens a feed, e.g. Subscribed feed. This should let
iOS count how often the user does this and if there is an appropriate widget on
the Home screen, then Siri will use this info to recommend the action and also
show our widget in the Siri smart stacks.