PostDetail: fixes a crash on comment reload
For some unknown to me reason, once pull-down-to-refresh is triggered and then
if app goes to background and we trigger context.safeIfNeeded() - this triggers
Core Data change notification that comes with an empty LemmyComment element. We
were crashing because the comment wasn't initialised with data and the "depth"
property was set to zero - which lead to colors[depth-1] triggering out of
bounds access.