Allow opening post links to any Lemmy instance
This allows to open PostDetail for any instance, not only for the currently
logged in instance. For example, I could be logged in to discuss.tchncs.de but
tap on a link
`info.ddenis.spud://internal/post?postId=1283466&instance=nerdly.dev` - this
creates a new PostDetail scene using signed out account and pushes that to the
navigation stack.

More UI work will be needed though, we should inform the user that they are
watching content from a foreign instance and that it's read-only. We should
also offer the user to load it from the home instance or an original instance
instead.

To make this complicated:
- User is logged in to discuss.tchncs.de
- User taps on a post https://nerdly.dev/post/1283466

We show PostDetail of the post in nerdly.dev instance, but this post is
originally from lemmy.world (https://lemmy.world/post/2846523).

Since this post is not on tchncs, will the user be able to comment/vote on it?