Avoid populating LemmyPerson.personInfo when only partial data is available
Previously when fetching a post we created a LemmyPerson and the corresponding
LemmyPersonInfo based on the data contained in the PostView - that one contains
LemmyKit.Person object with much (but not all) the info about the person.

However since now we display PersonVC based on LemmyPersonInfo, it is ambiguous
if the personInfo contains full data or only partial data.

The solution is to ensure that LemmyPerson.personInfo is only populated when we
fetch full info about the person (getPersonDetails call), but to ensure we have
the basic info to display in the post list - store partial data in LemmyPerson
itself as a cached/ephemeral value.