Added LemmyFeed.identifierForDebugging
We get a LemmyPost duplicate - assertion in LemmyPost.upsert has been triggered
- we have two posts with the same postId for the same account. The assumption
is that this is due to merging of changes coming from different Core Data
contexts - maybe a widget was fetching feed at the same time as the app was
fetching a feed.

still looking for a good solution for this (there are some challenges to add
uniqueness constraints), but for now lets try to debug this.