Extracted LemmyDataService from LemmyService
Having LemmyService be both a place for making api calls to remote servers and
also managing local CoreData types was problematic when it came for strict
concurrency. On one side the api calls could be made from any thread so it
makes sense for LemmyService to be an actor, but on the other side core data
management needs to happen on the main thread in the mainContext.