Quick fix for strict concurrency warnings
AccountService should probably not be bound to main thread as we want to use it
from SchedulerService (which in the future will be running in background
thread). And currently the functions in AccountService take ManagedContext as
an argument which won't always be main context. So this is a quick hack, need
to think more about it.